refactor: inline single-use workflow step helpers
Keep helper code inside the owning step when it is not shared across steps. Inline the archive and archive-tier helper logic into their step files and keep RestoreStepSupport as the remaining shared workflow helper.
